Add `description` field to `alarms` examples (#187)
In trying to diagnose an issue with why our ICS files were not being
validated correctly on iCalendar Validator due to a missing
'DESCRIPTION' within 'VALARM', I came across the ticket
from YM-KA stating that you could add the description field to the
alarms object without issue. I ran across in an iCalendar spec that the
description is one of three required fields for it to be valid as well,
so I thought that the README and it's examples regarding alarms should
reflect that, as there is currently no documentation of the description
field at all.  This would have saved me some time and hopefully will save
others the same fruustration.

This PR adds a `description` field to all instances of the `alarms`
array and the objects contained within.